[elsa] readOnly method add conditionNode logic
